pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"


[project]
name = "lah"
description="A package for learning algorithm steps, using data."
readme = "README.md"
license = {text = "Apache 2.0"}
dynamic = ["version"]
authors = [
{name = "Rajiv Sambharya", email="sambhar9@seas.upenn.edu"},
{name = "Bartolomeo Stellato", email="bstellato@princeton.edu"}
]
dependencies = [
    "numpy",
    "scipy==1.12",
    "jax==0.4.23",
    "jaxlib==0.4.23",
    "cvxpy>=1.3.0",
    "matplotlib",
    "jaxopt==0.8",
    "optax==0.1.5",
    "matplotlib",
    "hydra-core", 
    "emnist", 
    "imageio",
    "networkx"
]

[tool.setuptools.packages.find]
include = ["lah*", "benchmarks*"]
exclude = ["tutorials*", "tests*"]

[tool.setuptools_scm]
# To infer version automatically from git
write_to = "lah/_version.py"

[project.optional-dependencies]
dev = ["pytest", "ruff", "ruff-lsp", "black", "pandas", "jupyterlab"]


[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
select = ["E", "F", "I"]
ignore = ["E722"]
line-length = 100
exclude = ["build", "examples", "instances", "docs", "*__init__.py"]
target-version = "py310"
